+++ build.sbt
@@ -3,7 +3,7 @@ name := "ideprobe-pants"
 organization.in(ThisBuild) := "com.twitter.ideprobe"
 version.in(ThisBuild) := "0.1"
 scalaVersion.in(ThisBuild) := "2.13.1"
-intellijBuild.in(ThisBuild) := "203.4203.26"
+intellijBuild.in(ThisBuild) := "203.6682.115"
 resolvers.in(ThisBuild) ++= Dependencies.ideProbe.resolvers
 skip in publish := true
